add options on services, session type

+++ src/modules/tags/tags.service.ts
@@ -1,6 +1,6 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model } from 'mongoose';
+import mongoose, { Model } from 'mongoose';
 import { CreateTagDto } from './dto/create-tag.dto';
 import { FindAllFilterTagDto } from './dto/find-all-filter-tag.dto';
 import { UpdateTagDto } from './dto/update-tag.dto';
@@ -41,11 +41,34 @@ export class TagsService {
     };
   }
 
-  async findOne(field: keyof Tag, fieldValue: unknown): Promise<Tag> {
+  async findOne(field: keyof Tag, fieldValue: unknown): Promise<Tag>;
+  async findOne(
+    field: keyof Tag,
+    fieldValue: unknown,
+    options: {
+      throw?: true;
+    },
+  ): Promise<Tag>;
+  async findOne(
+    field: keyof Tag,
+    fieldValue: unknown,
+    options: {
+      throw?: false;
+    },
+  ): Promise<Tag | null>;
+  async findOne(
+    field: keyof Tag,
+    fieldValue: unknown,
+    options: {
+      throw?: boolean;
+    } = { throw: true },
+  ): Promise<Tag | null> {
     let foundTag: Tag | null = null;
     switch (field) {
       case '_id': {
-        foundTag = await this.tagModel.findOne({ _id: fieldValue }).exec();
+        if (mongoose.Types.ObjectId.isValid(fieldValue as string)) {
+          foundTag = await this.tagModel.findOne({ _id: fieldValue }).exec();
+        }
         break;
       }
       case 'slug': {
@@ -56,18 +79,52 @@ export class TagsService {
         break;
       }
     }
-    if (!foundTag) {
+    if (!foundTag && options.throw) {
       throw new NotFoundException('Tag Not Found');
     }
     return foundTag;
   }
 
-  async updateOne(field: keyof Tag, fieldValue: unknown, updateDto: Partial<UpdateTagDto>): Promise<Tag> {
+  async updateOne(field: keyof Tag, fieldValue: unknown, updateDto: Partial<UpdateTagDto>): Promise<Tag>;
+  async updateOne(
+    field: keyof Tag,
+    fieldValue: unknown,
+    updateDto: Partial<UpdateTagDto>,
+    options: {
+      throw?: true;
+    },
+  ): Promise<Tag>;
+  async updateOne(
+    field: keyof Tag,
+    fieldValue: unknown,
+    updateDto: Partial<UpdateTagDto>,
+    options: {
+      throw?: false;
+    },
+  ): Promise<Tag | null>;
+  async updateOne(
+    field: keyof Tag,
+    fieldValue: unknown,
+    updateDto: Partial<UpdateTagDto>,
+    options: {
+      throw?: boolean;
+    } = { throw: true },
+  ): Promise<Tag | null> {
     let updatedTag: Tag | null = null;
     switch (field) {
       case '_id': {
+        if (mongoose.Types.ObjectId.isValid(fieldValue as string)) {
+          updatedTag = await this.tagModel
+            .findOneAndUpdate({ _id: fieldValue }, updateDto, {
+              new: true,
+            })
+            .exec();
+        }
+        break;
+      }
+      case 'slug': {
         updatedTag = await this.tagModel
-          .findOneAndUpdate({ _id: fieldValue }, updateDto, {
+          .findOneAndUpdate({ slug: fieldValue }, updateDto, {
             new: true,
           })
           .exec();
@@ -77,24 +134,54 @@ export class TagsService {
         break;
       }
     }
-    if (!updatedTag) {
+    if (!updatedTag && options.throw) {
       throw new NotFoundException('Tag Not Updated');
     }
     return updatedTag;
   }
 
-  async deleteOne(field: keyof Tag, fieldValue: unknown): Promise<Tag> {
+  async deleteOne(field: keyof Tag, fieldValue: unknown): Promise<Tag>;
+  async deleteOne(
+    field: keyof Tag,
+    fieldValue: unknown,
+    options: {
+      secret?: boolean;
+      throw?: true;
+    },
+  ): Promise<Tag>;
+  async deleteOne(
+    field: keyof Tag,
+    fieldValue: unknown,
+    options: {
+      secret?: boolean;
+      throw?: false;
+    },
+  ): Promise<Tag | null>;
+  async deleteOne(
+    field: keyof Tag,
+    fieldValue: unknown,
+    options: {
+      secret?: boolean;
+      throw?: boolean;
+    } = { secret: false, throw: true },
+  ): Promise<Tag | null> {
     let deletedTag: Tag | null = null;
     switch (field) {
       case '_id': {
-        deletedTag = await this.tagModel.findByIdAndRemove({ _id: fieldValue }).exec();
+        if (mongoose.Types.ObjectId.isValid(fieldValue as string)) {
+          deletedTag = await this.tagModel.findByIdAndRemove({ _id: fieldValue }).exec();
+        }
+        break;
+      }
+      case 'slug': {
+        deletedTag = await this.tagModel.findByIdAndRemove({ slug: fieldValue }).exec();
         break;
       }
       default: {
         break;
       }
     }
-    if (!deletedTag) {
+    if (!deletedTag && options.throw) {
       throw new NotFoundException('Tag Not Deleted');
     }
     return deletedTag;
